Fixing a boatload of SA and CA warnings and decided to introduce Next/Previous navigators for workitem revisions.
