+++ Aggregator.Core/Facade/WorkItemWrapper.cs
@@ -171,13 +171,37 @@ namespace Aggregator.Core.Facade
         {
             get
             {
-                // Revision starts at 1, so Revision-1 is last and Revision-2 is next-to-last
-                int lastRevisionNumber = this.workItem.Revision - 2;
-                if (lastRevisionNumber < 0)
+                return new RevisionWrapper(this.workItem.Revisions[this.workItem.Revisions.Count - 1]);
+            }
+        }
+
+        public IRevision PreviousRevision
+        {
+            get
+            {
+                int targetRevision = this.workItem.Revision - 1;
+
+                if (targetRevision <= 1)
                 {
-                    lastRevisionNumber = 0;
+                    targetRevision = 1;
                 }
-                return new RevisionWrapper(this.workItem.Revisions[lastRevisionNumber]);
+
+                return new RevisionWrapper(this.workItem.Revisions[targetRevision - 1]);
+            }
+        }
+
+        public IRevision NextRevision
+        {
+            get
+            {
+                int targetRevision = this.workItem.Revision + 1;
+
+                if (targetRevision <= this.workItem.Revisions.Count + 1)
+                {
+                    targetRevision = this.workItem.Revisions.Count - 1;
+                }
+
+                return new RevisionWrapper(this.workItem.Revisions[targetRevision]);
             }
         }
