+++ Aggregator.Core/Monitoring/ITextLogger.cs
@@ -3,6 +3,7 @@
     public interface ITextLogger : ILogger
     {
         void Log(LogLevel level, string format, params object[] args);
+
         void UserLog(LogLevel level, string ruleName, string message);
     }
 }
